services/terra-data-collector/requirements.txt
# ============================
# ⚙️ Core Framework
# ============================
fastapi>=0.109.0                 # API server for health/status endpoints
uvicorn[standard]>=0.27.0        # ASGI server
pydantic>=2.5.3                  # Data validation & models
pydantic-settings>=2.1.0         # Environment variable management
pyyaml>=6.0.1                    # YAML config parsing

# ============================
# 🌐 Data Collection
# ============================
httpx>=0.27.0                    # Async HTTP client for API calls
beautifulsoup4>=4.12.2           # HTML parsing for web scraping fallback

# ============================
# 🔄 Messaging (optional)
# ============================
# aiokafka>=0.10.0              # Direct Kafka bridge (uncomment if needed)

# ============================
# 📊 Data Export & Processing
# ============================
pandas>=2.2.0                    # DataFrame operations for export
pyarrow>=15.0.0                  # Parquet export support

# ============================
# ⏰ Scheduling
# ============================
apscheduler>=3.10.4              # Task scheduling

# ============================
# 📈 Monitoring
# ============================
prometheus-client>=0.19.0        # Prometheus metrics

# ============================
# 🛠️ CLI
# ============================
click>=8.1.7                     # CLI framework
rich>=13.7.0                     # Rich terminal output
